Refactored schema/typeorm-schema to smaller files
