+++ lib/schema/typeorm-schema/index.ts
@@ -1,0 +1,6 @@
+export * from './typeorm-brackets'
+export * from './typeorm-column-info'
+export * from './typeorm-query-builder'
+export * from './typeorm-select-query-builder'
+export * from './typeorm-table-info'
+export * from './types'
