+++ lib/schema/typeorm-schema/typeorm-brackets.ts
@@ -1,0 +1,6 @@
+import { Brackets } from 'typeorm'
+import { IBrackets } from '../types'
+
+export class TypeOrmBrackets implements IBrackets {
+    constructor(public readonly data: Brackets) {}
+}
