+++ lib/schema/typeorm-schema/typeorm-select-query-builder.ts
@@ -1,0 +1,6 @@
+import { SelectQueryBuilder } from 'typeorm'
+import { TypeOrmQueryBuilder } from './typeorm-query-builder'
+
+export class TypeOrmSelectQueryBuilder extends TypeOrmQueryBuilder {
+    declare data: SelectQueryBuilder<any>
+}
