+++ lib/schema/typeorm-schema/types.test.ts
@@ -1,0 +1,7 @@
+import { Repository } from 'typeorm'
+import { Book, TestDatabase } from '@/test-db'
+
+export interface TypeOrmSchema_TestContext {
+    db: TestDatabase
+    repo: Repository<Book>
+}
