+++ tsconfig.base.json
@@ -27,6 +27,6 @@
     "exclude": [
         "node_modules",
         "./dist/**/*",
-        "./test/**/*"
+        "./test/**/*",
     ]
 }
